+++ src/lib/components/upload/UploadSection.tsx
@@ -70,7 +70,7 @@ export const UploadSection = ({
         onTxSucceed: (codeId: number) => {
           updateCodeInfo(
             codeId,
-            address,
+            address as HumanAddr,
             codeDesc || `${wasmFile?.name}(${codeId})`
           );
         },