+++ src/lib/gql/graphql.ts
@@ -6657,16 +6657,14 @@ export type GetExecuteTxsByContractAddressQueryVariables = Exact<{
 
 export type GetExecuteTxsByContractAddressQuery = {
   __typename?: "query_root";
-  contract_transactions: Array<{
-    __typename?: "contract_transactions";
-    transaction: {
-      __typename?: "transactions";
-      hash: any;
-      messages: any;
-      success: boolean;
-      account: { __typename?: "accounts"; address: string };
-      block: { __typename?: "blocks"; height: number; timestamp: any };
-    };
+  contract_transactions_view: Array<{
+    __typename?: "contract_transactions_view";
+    hash?: any | null;
+    messages?: any | null;
+    success?: boolean | null;
+    sender?: string | null;
+    height?: number | null;
+    timestamp?: any | null;
   }>;
 };
 
@@ -7906,7 +7904,7 @@ export const GetExecuteTxsByContractAddressDocument = {
         selections: [
           {
             kind: "Field",
-            name: { kind: "Name", value: "contract_transactions" },
+            name: { kind: "Name", value: "contract_transactions_view" },
             arguments: [
               {
                 kind: "Argument",
@@ -7916,28 +7914,16 @@ export const GetExecuteTxsByContractAddressDocument = {
                   fields: [
                     {
                       kind: "ObjectField",
-                      name: { kind: "Name", value: "contract" },
+                      name: { kind: "Name", value: "contract_address" },
                       value: {
                         kind: "ObjectValue",
                         fields: [
                           {
                             kind: "ObjectField",
-                            name: { kind: "Name", value: "address" },
+                            name: { kind: "Name", value: "_eq" },
                             value: {
-                              kind: "ObjectValue",
-                              fields: [
-                                {
-                                  kind: "ObjectField",
-                                  name: { kind: "Name", value: "_eq" },
-                                  value: {
-                                    kind: "Variable",
-                                    name: {
-                                      kind: "Name",
-                                      value: "contractAddress",
-                                    },
-                                  },
-                                },
-                              ],
+                              kind: "Variable",
+                              name: { kind: "Name", value: "contractAddress" },
                             },
                           },
                         ],
@@ -7945,23 +7931,14 @@ export const GetExecuteTxsByContractAddressDocument = {
                     },
                     {
                       kind: "ObjectField",
-                      name: { kind: "Name", value: "transaction" },
+                      name: { kind: "Name", value: "is_execute" },
                       value: {
                         kind: "ObjectValue",
                         fields: [
                           {
                             kind: "ObjectField",
-                            name: { kind: "Name", value: "is_execute" },
-                            value: {
-                              kind: "ObjectValue",
-                              fields: [
-                                {
-                                  kind: "ObjectField",
-                                  name: { kind: "Name", value: "_eq" },
-                                  value: { kind: "BooleanValue", value: true },
-                                },
-                              ],
-                            },
+                            name: { kind: "Name", value: "_eq" },
+                            value: { kind: "BooleanValue", value: true },
                           },
                         ],
                       },
@@ -7977,26 +7954,8 @@ export const GetExecuteTxsByContractAddressDocument = {
                   fields: [
                     {
                       kind: "ObjectField",
-                      name: { kind: "Name", value: "transaction" },
-                      value: {
-                        kind: "ObjectValue",
-                        fields: [
-                          {
-                            kind: "ObjectField",
-                            name: { kind: "Name", value: "block" },
-                            value: {
-                              kind: "ObjectValue",
-                              fields: [
-                                {
-                                  kind: "ObjectField",
-                                  name: { kind: "Name", value: "timestamp" },
-                                  value: { kind: "EnumValue", value: "desc" },
-                                },
-                              ],
-                            },
-                          },
-                        ],
-                      },
+                      name: { kind: "Name", value: "timestamp" },
+                      value: { kind: "EnumValue", value: "desc" },
                     },
                   ],
                 },
@@ -8021,54 +7980,12 @@ export const GetExecuteTxsByContractAddressDocument = {
             selectionSet: {
               kind: "SelectionSet",
               selections: [
-                {
-                  kind: "Field",
-                  name: { kind: "Name", value: "transaction" },
-                  selectionSet: {
-                    kind: "SelectionSet",
-                    selections: [
-                      { kind: "Field", name: { kind: "Name", value: "hash" } },
-                      {
-                        kind: "Field",
-                        name: { kind: "Name", value: "messages" },
-                      },
-                      {
-                        kind: "Field",
-                        name: { kind: "Name", value: "success" },
-                      },
-                      {
-                        kind: "Field",
-                        name: { kind: "Name", value: "account" },
-                        selectionSet: {
-                          kind: "SelectionSet",
-                          selections: [
-                            {
-                              kind: "Field",
-                              name: { kind: "Name", value: "address" },
-                            },
-                          ],
-                        },
-                      },
-                      {
-                        kind: "Field",
-                        name: { kind: "Name", value: "block" },
-                        selectionSet: {
-                          kind: "SelectionSet",
-                          selections: [
-                            {
-                              kind: "Field",
-                              name: { kind: "Name", value: "height" },
-                            },
-                            {
-                              kind: "Field",
-                              name: { kind: "Name", value: "timestamp" },
-                            },
-                          ],
-                        },
-                      },
-                    ],
-                  },
-                },
+                { kind: "Field", name: { kind: "Name", value: "hash" } },
+                { kind: "Field", name: { kind: "Name", value: "messages" } },
+                { kind: "Field", name: { kind: "Name", value: "success" } },
+                { kind: "Field", name: { kind: "Name", value: "sender" } },
+                { kind: "Field", name: { kind: "Name", value: "height" } },
+                { kind: "Field", name: { kind: "Name", value: "timestamp" } },
               ],
             },
           },