+++ src/lib/model/code.ts
@@ -32,7 +32,7 @@ export const useCodeContractInstances = (
   codeId: number,
   offset: number,
   pageSize: number
-): Option<ContractInstances> => {
+): ContractInstances => {
   const { data: contractList } = useContractListByCodeId(
     codeId,
     offset,