+++ src/lib/pages/contract-details/components/tables/execute/ExecuteTableRow.tsx
@@ -60,10 +60,16 @@ export const ExecuteTableRow = ({
       </TableRow>
       <TableRow>
         <Flex direction="column" gap={1}>
-          <Text variant="body3">{formatUTC(transaction.created)}</Text>
-          <Text variant="body3" color="text.dark">
-            {`(${dateFromNow(transaction.created)})`}
-          </Text>
+          {transaction.created ? (
+            <>
+              <Text variant="body3">{formatUTC(transaction.created)}</Text>
+              <Text variant="body3" color="text.dark">
+                {`(${dateFromNow(transaction.created)})`}
+              </Text>
+            </>
+          ) : (
+            <Text variant="body3">N/A</Text>
+          )}
         </Flex>
       </TableRow>
     </Grid>