+++ src/lib/types/currency/index.ts
@@ -1,4 +1,3 @@
-export * from "../addrs";
 export * from "./common";
 export * from "./token";
 export * from "./balance";